+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### ✨ Improved
+
+* Add Overwatcher daily task to power cycle AG cameras before observations.
+
+
 ## 1.7.2 - March 2, 2025
 
 ### ✨ Improved